[feature] complete detail of movie screen / complete comment section in detail movie / add read only option to artist_section, country_multi_selector, date_picker_section, genre_section, poster_section, synopsis_section, thumbnail_section, title_section and value_section widget in movie screen

+++ lib/feature/movie/presentation/widget/poster_section_widget/poster_section_widget.dart
@@ -5,7 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class PosterSectionWidget extends StatelessWidget {
-  const PosterSectionWidget({super.key});
+  final bool readOnly;
+
+  const PosterSectionWidget({super.key, this.readOnly = false});
 
   @override
   Widget build(BuildContext context) {
@@ -127,11 +129,13 @@ class PosterSectionWidget extends StatelessWidget {
               overflow: TextOverflow.ellipsis),
         ),
         const SizedBox(width: 8),
-        IconButton(
-            onPressed: () {
-              BlocProvider.of<PosterSectionCubit>(context).removeFile();
-            },
-            icon: const Icon(Icons.close, color: Colors.red))
+        if (!readOnly) ...[
+          IconButton(
+              onPressed: () {
+                BlocProvider.of<PosterSectionCubit>(context).removeFile();
+              },
+              icon: const Icon(Icons.close, color: Colors.red))
+        ]
       ],
     );
   }
